Sample Dataset for CyTOF Stimulation Assay

A list with the CyTOF stimulation assay data.

chi11_1k

Format

A list with one tibble containig CyTOF expression data. And four character vectors for arguments in the stim_cell_selector function.

chi11_1k$expr_data

A 70,000 X 36 tibble. Cells are on the rows and variables on the columns. The first 6 columns contain for each cell cluster_id (from FlowSOM clustering), sample_id (unique for each FSC file), condition (comparison groups), patient_id (unique for each subject), stim_type (labels for types of stimulation assays including the unstim), merging1 (meta culster labels from ConsensusClusterPlus). The last 30 columns contain the archsinh transformed CyTOF expression values for the 30 markers (20 type and 10 state) used in the sitmulation panel.

chi11_1k$type_markers

A character vector with the labels for type markers used in the stimulation panel.

chi11_1k$state_markers

A character vector with the labels for state markers used in the stimulation panel.

chi11_1k$cluster_col

A character label of the meta-cluster/cluster ID column in chi11_1k$expr_dat tibble.

chi11_1k$stim_label

A character vector with the label(s) for the stimulation types corresponding to the labels in thestim_type column in chi11_1k$expr_data.

chi11_1k$unstim_label

A character label for the unstim cells corresponding to the labels in thestim_type column in chi11_1k$expr_data.