Stim Cell Selector Single Marker

Function to select cells from the stimulated samples that have likely responded to the stimulant.

stim_cell_selector_single_marker(dat, state_markers, cluster_col, stim_lab,
  unstim_lab, seed_val = NULL, verbose = FALSE)

Arguments

dat
A tibble with the single cell data. Cells on rows and variables/markers on columns.

state_markers
A character vector with the labels of state markers from the stimulation panel.

cluster_col
Column in the tibble with the cluster IDs.

stim_lab
A character of stim label(s).

unstim_lab
A character of unstim label(s).

seed_val
Seed value (integer) for kmeans clustering. Default is NULL for no seed value.

verbose
Logical. To make function more verbose. Default is FALSE.

Value

A list with tibbles for expression data for the selected cells, data to plot stacked bar plots, data to plot UMAP plots, and parameters passed to the function.

Examples

selected_data <- stim_cell_selector_single_marker(chi11_1k$expr_data, chi11_1k$state_markers,
                  chi11_1k$cluster_col, chi11_1k$stim_label,
                  chi11_1k$unstim_label, seed_val = 123, verbose = FALSE)

#> 62386337

#> U

#> 78228317

#> U

#> 1189991

#> U

#> 3961496

#> U

#> 88129010

#> U

#> 2179298

#> U

#> 37634060

#> U

#> 44564753

#> U

#> 964955

#> U

#> 919937

#> U

#> 73277327

#> U

#> 31698317

#> U

#> 13871189

#> U

#> 19818911

#> U

#> 81198218

#> U

#> 4852937

#> U

#> 11891387

#> U

#> 40603961

#> U

#> 17831783

#> U

#> 13871486

#> U

#> 78228020

#> U

#> 60401783

#> U

#> 88129010

#> U

#> 67331585

#> U

#> 86148416

#> U

#> 6040973

#> U

#> 25753070

#> U

#> 42584654

#> U

#> 11891288

#> U

#> 88128812

#> U

#> 36643862

#> U

#> 66341882

#> U

#> 88128812

#> U

#> 79211189

#> U

#> 82188317

#> U

#> 50501189

#> U

#> 85158515

#> U

#> 43574555

#> U

#> 79218119

#> U

#> 10901090

#> U

#> 31693169

#> U

#> 58421981

#> U

#> 87138911

#> U

#> 7525694

#> U

#> 19811882

#> U

#> 54462080

#> U

#> 18821585

#> U

#> 51495149

#> U

#> 76247624

#> U

#> 9010919

#> U

#> 30702872

#> U

#> 50501486

#> U

#> 928928

#> U

#> 62381486

#> U

#> 13871288

#> U

#> 3169496

#> U

#> 66346733

#> U

#> 47534456

#> U

#> 85158416

#> U

#> 12881288

#> U

#> 36643565

#> U

#> 42581981

#> U

#> 19811882

#> U

#> 6733892

#> U

#> 21791981

#> U

#> 64362575

#> U

#> 67336436

#> U

#> 50504852

#> U

#> 8921189

#> U

#> 60406040

#> U

#> 68326832

#> U

#> 65358317

#> U

#> 80208020

#> U

#> 5842793

#> U

#> 13871090

#> U

#> 53471387

#> U

#> 63376040

#> U

#> 53475347

#> U

#> 955964

#> U

#> 60405941

#> U

#> 33673169

#> U

#> 45557525

#> U

#> 16841486

#> U

#> 9918317

#> U

#> 39613466

#> U

#> 6733595

#> U

#> 56445545

#> U

#> 65356337

#> U

#> 892892

#> U

#> 18821783

#> U

#> 85158713

#> U

#> 15851090

#> U

#> 928919

#> U

#> 991298

#> U

#> 28723070

#> U

#> 991397

#> U

#> 31693070

#> U

#> 43573961

#> U

#> 694595

#> U

#> 88128911

#> U

#> 65356931

#> U

#> 34662080

#> U

#> 14861387

#> U

#> 57432080

#> U

#> 71297921

#> U

#> 3169595

#> U

#> 32683763

#> U

#> 34662674

#> U

#> 955946

#> U

#> 86148218

#> U

#> 12882674

#> U

#> 41591783

#> U

#> 24763070

#> U

#> 47530100

#> U

#> 82187426

#> U

#> 7624919

#> U

#> 47536535

#> U

#> 53475743

#> U

#> 8812919

#> U

#> 946919

#> U

#> 48523466

#> U

#> 26741288

#> U

#> 11891486

#> U

#> 56441189

#> U

#> 81198911

#> U

#> 7723964

#> U

#> 60405743

#> U

#> 39614555

#> U

#> 694793

#> U

#> 74268020

#> U

#> 15851189

#> U

#> 62388317

#> U

#> 6941783

#> U

#> 47530100

#> U

#> 22782377

#> U

#> 69318812

#> U

#> 45555248

#> U

#> 40603565

#> U

#> 87138614

#> U

#> 8713955

#> U

#> 73277723

#> U

#> 56441981

#> U

#> 13871189

#> U

#> 7426496

#> U

#> 20801189

#> U

#> 54461486

#> U

#> 84168713

#> U

#> 50504951

#> U

#> 28722377

#> U

#> 13871189

#> U

#> 73277624

#> U

#> 39618416

#> U

#> 496793

#> U

#> 3466937

#> U

#> 77238317

#> U

#> 5941973

#> U

#> 23773268

#> U

#> 43575050

#> U

#> 84168713

#> U

#> 496595

#> U

#> 9911000

#> U

#> 23772278

#> U

#> 928919

#> U

#> 19811882

#> U

#> 38623367

#> U

#> 919964

#> U

#> 10901684

#> U

#> 68325644

#> U

#> 946937

#> U

#> 14861288

#> U

#> 62385842

#> U

#> 51496931

#> U

#> 8614928

#> U

#> 22787525

#> U

#> 79218020

#> U

#> 3961298

#> U

#> 47535644

#> U

#> 21792773

#> U

#> 10901486

#> U

#> 81197228

#> U

#> 75258317

#> U

#> 5545937

#> U

#> 928937

#> U

#> 5545937

#> U

#> 18821684

#> U

#> 6535955

#> U

#> 50505941

#> U

#> 68328119

#> U

#> 90108614

#> U

#> 10901486

#> U

#> 29713961

#> U

#> 42583961

#> U

#> 9468515

#> U

#> 4852397

#> U

#> 9467921

#> U

#> 7129946

#> U

#> 32684258

#> U

#> 48525545

#> U

#> 694694

#> U

#> 9010919

#> U

#> 34663763

#> U

#> 18821882

#> U

#> 8812919

#> U

#> 90109010

#> U

#> 10901090

#> U

#> 955946

#> U

#> 59416139

#> U

#> 53475347

#> U

#> 964955

#> U

#> 955946

#> U

#> 35652476

#> U

#> 82188020

#> U

#> 6436496

#> U

#> 42582971

#> U

#> 55458812

#> U

#> 2575991

#> U

#> 12881288

#> U

#> 30704555

#> U

#> 4852892

#> U

#> 79219010

#> U

#> 73278119

#> U

#> 82188218

#> U

#> 4357973

#> U

#> 33672872

#> U

#> 65359010

#> U

#> 8020964

#> U

#> 31693070

#> U

#> 38624951

#> U

#> 5941946

#> U

#> 83178812

#> U

#> 54466733

#> U

#> 20802080

#> U

#> 3169973

#> U

#> 43572971

#> U

#> 52488911

#> U

#> 67338713

#> U

#> 16841486

#> U

#> 28725248

#> U

#> 5347991

#> U

#> 1486991

#> U

#> 56447228

#> U

#> 17832080

#> U

#> 2575982

#> U

#> 60408317

#> U

#> 47539010

#> U

#> 59417822

#> U

#> 17831486

#> U

#> 24766139

#> U

#> 38628812

#> U

#> 8713928

#> U

#> 35652872

#> U

#> 84168416

#> U

#> 3862973

#> U

#> 39612971

#> U

#> 4159991

#> U

#> 7624946

#> U

#> 62386832

#> U

#> 55454060

#> U

#> 5842955

#> U

#> 8911955

#> U

#> 34663367

#> U

#> 84168119

#> U

#> 7228946

#> U

#> 78228020

#> U

#> 76248911

#> U

#> 66346931

#> U

#> 40603268

#> U

#> 60404555

#> U

#> 1585298

#> U

#> 40604555

#> U

#> 67336931

#> U

#> 85158317

#> U

#> 8218982

#> U

#> 78228812

#> U

#> 8515964

#> U

#> 21791486

#> U

#> 40603862

#> U

#> 47535644

#> U

#> 86141000

#> U

#> 40604258

#> U

#> 31693070

#> U

#> 30702872

#> U

#> 64369010

#> U

#> 56444555

#> U

#> 52486931

#> U

#> 77237822

#> U

#> 42584951

#> U

#> 24764555

#> U

#> 8119946

#> U

#> 7327928

#> U

#> 13871288

#> U

#> 87139010

#> U

#> 2278694

#> U

#> 928928

#> U

#> 70307129

#> U

#> 9010937

#> U

#> 30703070

#> U

#> 41594060

#> U

#> 8713955

#> U

#> 87139010

#> U

#> 66346634

#> U

#> 14862278

#> U

#> 21791387

#> U

#> 26742773

#> U

#> 28722080

#> U

#> 88128515

#> U

#> 58426337

#> U

#> 64367426

#> U

#> 1684694

#> U

#> 83178515

#> U

#> 12882674

#> U

#> 27731783

#> U

#> 38620100

#> U

#> 35652278

#> U

#> 27732278

#> U

#> 496991

#> U

#> 46547426

#> U

#> 35655743

#> U

#> 4258991

#> U

#> 8119919

#> U

#> 42583862

#> U

#> 76248812

#> U

#> 18822080

#> U

#> 35653466

#> U

#> 15851288

#> U

#> 694793

#> U

#> 56445743

#> U

#> 45555545

#> U

#> 1090793

#> U

#> 68327921

#> U

#> 14861585

#> U

#> 23772080

#> U

#> 595298

#> U

#> 25752971

#> U

#> 1189892

#> U

#> 23772179

#> U

#> 52484753

#> U

#> 35653070

#> U

#> 6940100

#> U

#> 892595

#> U

#> 67338020

#> U

#> 16841981

#> U

#> 6733496

#> U

#> 39612476

#> U

#> 4159793

#> U

#> 24761882

#> U

#> 14861387

#> U

#> 26745545

#> U

#> 4852694

#> U

#> 19811090

#> U

#> 31692278

#> U

#> 85158416

#> U

#> 4852982

#> U

#> 36642773

#> U

#> 66348911

#> U

#> 7723955

#> U

#> 67336832

#> U

#> 59414852

#> U

#> 5842973

#> U

#> 7921946

#> U

#> 9010937

#> U

#> 80207723

#> U

#> 2872595

#> U

#> 85158812

#> U

#> 66346733

#> U

#> 1288595

#> U

#> 78228218

#> U

#> 38624654

#> U

#> 1486199

#> U

#> 16841288

#> U

#> 60405644

#> U

#> 16843367

#> U

#> 8416919

#> U

#> 41594456

#> U

#> 20801981

#> U

#> 8614928

#> U

#> 56444555

#> U

#> 26742773

#> U

#> 7931684

#> U

#> 82188119

#> U

#> 18821783

#> U

#> 9010919

#> U

#> 3268982

#> U

#> 82188614

#> U

#> 60408812

#> U

#> 1882793

#> U

#> 54465545

#> U

#> 74268119

#> U

#> 38621486

#> U

#> 82188812

#> U

#> 44566139

#> U

#> 23773961

#> U

#> 4961288

#> U

#> 1783991

#> U

#> 90108515

#> U

#> 964919

#> U

#> 71298218

#> U

#> 50505545

#> U

#> 694694

#> U

#> 928919

#> U

#> 35653664

#> U

#> 17831882

#> U

#> 919919

#> U

#> 11891189

#> U

#> 11891090

#> U

#> 595595

#> U

#> 60405941

#> U

#> 56445347

#> U

#> 955955

#> U

#> 928928

#> U

#> 25752773

#> U

#> 80208020

#> U

#> 88128812

#> U

#> 71296832

#> U

#> 19811882

#> U

#> 10901189

#> U

#> 14861387

#> U

#> 62386139

#> U

#> 19811981

#> U

#> 86148713

#> U

#> 22782179

#> U

#> 17831981

#> U

#> 9911090

#> U

#> 25753070

#> U

#> 14861486

#> U

#> 955955

#> U

#> 72287030

#> U

#> 55455644

#> U

#> 85158713

#> U

#> 13871486

#> U

#> 64366337

#> U

#> 80208020

#> U

#> 88128812

#> U

#> 31693367

#> U

#> 18821882

#> U

#> 14861585

#> U

#> 86148614

#> U

#> 57435644

#> U

#> 21792080

#> U

#> 10901090

#> U

#> 71297030

#> U

#> 83178119

#> U

#> 89118911

#> U

#> 20802080

#> U

#> 83178218

#> U

#> 81197723

#> U

#> 15851486

#> U

#> 54465050

#> U

#> 76247624

#> U

#> 1090991

#> U

#> 70307228

#> U

#> 84168416

#> U

#> 90108812

#> U

#> 74267030

#> U

#> 13871288

#> U

#> 946946

#> U

#> 33673268

#> U

#> 55455644

#> U

#> 85158416

#> U

#> 11891288

#> U

#> 34663664

#> U

#> 15851981

#> U

#> 78227723

#> U

#> 18822377

#> U

#> 79218020

#> U

#> 30703268

#> U

#> 64366634

#> U

#> 51495050

#> U

#> 7931090

#> U

#> 57436238

#> U

#> 68326832

#> U

#> 83178317

#> U

#> 80207921

#> U

#> 86148416

#> U

#> 13871189

#> U

#> 79217822

#> U

#> 61396040

#> U

#> 54465347

#> U

#> 964964

#> U

#> 54466040

#> U

#> 68327030

#> U

#> 22782872

#> U

#> 25751387

#> U

#> 46545347

#> U

#> 8614298

#> U

#> 28722179

#> U

#> 16846931

#> U

#> 83175842

#> U

#> 3466982

#> U

#> 2674595

#> U

#> 14861585

#> U

#> 9289010

#> U

#> 793892

#> U

#> 10901189

#> U

#> 29712971

#> U

#> 928937

#> U

#> 71297030

#> U

#> 57436238

#> U

#> 595397

#> U

#> 89119010

#> U

#> 76246634

#> U

#> 90107822

#> U

#> 86148713

#> U

#> 75256832

#> U

#> 6139955

#> U

#> 12881387

#> U

#> 73273664

#> U

#> 41592674

#> U

#> 73271000

#> U

#> 73279010

#> U

#> 24762674

#> U

#> 12881783

#> U

#> 4960100

#> U

#> 2080991

#> U

#> 8922278

#> U

#> 892991

#> U

#> 56447426

#> U

#> 56444357

#> U

#> 32681387

#> U

#> 1000919

#> U

#> 30703466

#> U

#> 84168812

#> U

#> 12882080

#> U

#> 35653466

#> U

#> 33671288

#> U

#> 793793

#> U

#> 49514357

#> U

#> 72284654

#> U

#> 991793

#> U

#> 35652080

#> U

#> 8921189

#> U

#> 71298020

#> U

#> 86148317

#> U

#> 27733070

#> U

#> 82187624

#> U

#> 87138218

#> U

#> 56445347

#> U

#> 28723565

#> U

#> 8921288

#> U

#> 8515955

#> U

#> 22782377

#> U

#> 16842080

#> U

#> 14861189

#> U

#> 78227228

#> U

#> 19811387

#> U

#> 87138218

#> U

#> 18821882

#> U

#> 59415149

#> U

#> 70307723

#> U

#> 13871189

#> U

#> 74267525

#> U

#> 79218416

#> U

#> 793694

#> U

#> 72287228

#> U

#> 81198317

#> U

#> 496397

#> U

#> 69316832

#> U

#> 42585050

#> U

#> 21791783

#> U

#> 946955

#> U

#> 964982

#> U

#> 10902278

#> U

#> 9468713

#> U

#> 82188218

#> U

#> 70306733

#> U

#> 964991

#> U

#> 74268218

#> U

#> 58425545

#> U

#> 937964

#> U

#> 82188416

#> U

#> 30704258

#> U

#> 18823367

#> U

#> 2476397

#> U

#> 56445842

#> U

#> 7624298

#> U

#> 928982

#> U

#> 76243961

#> U

#> 74264456

#> U

#> 68320100

#> U

#> 54460100

#> U

#> 71297624

#> U

#> 80208317

#> U

#> 1387793

#> U

#> 9911486

#> U

#> 9648812

#> U

#> 964937

#> U

#> 47535941

#> U

#> 53478119

#> U

#> 9911486

#> U

#> 18821288

#> U

#> 62386139

#> U

#> 75256139

#> U

#> 90108515

#> U

#> 892991

#> U

#> 10902179

#> U

#> 1000919

#> U

#> 85158218

#> U

#> 62385545

#> U

#> 595694

#> U

#> 892694

#> U

#> 64366337

#> U

#> 18821882

#> U

#> 9010928

#> U

#> 9010928

#> U

#> 11891090

#> U

#> 946964

#> U

#> 62386139

#> U

#> 47534753

#> U

#> 496595

#> U

#> 937937

#> U

#> 72287327

#> U

#> 80208020

#> U

#> 11891288

#> U

#> 60407228

#> U

#> 83178218

#> U

#> 10901090

#> U

#> 11891387

#> U

#> 40604060

#> U

#> 83178317

#> U

#> 14861387

#> U

#> 79218119

#> U

#> 18821882

#> U

#> 9289010

#> U

#> 64367228

#> U

#> 86148812

#> U

#> 928964

#> U

#> 27733169

#> U

#> 42584654

#> U

#> 11891288

#> U

#> 82188317

#> U

#> 35653763

#> U

#> 79218020

#> U

#> 88128812

#> U

#> 45552674

#> U

#> 80208218

#> U

#> 86148515

#> U

#> 85158515

#> U

#> 57435545

#> U

#> 20802080

#> U

#> 11891189

#> U

#> 68326931

#> U

#> 80208020

#> U

#> 10901189

#> U

#> 45551189

#> U

#> 80208119

#> U

#> 23772377

#> U

#> 15851585

#> U

#> 46545050

#> U

#> 75257624

#> U

#> 9010919

#> U

#> 32682971

#> U

#> 21791684

#> U

#> 89118812

#> U

#> 36642674

#> U

#> 12881288

#> U

#> 892694

#> U

#> 29713367

#> U

#> 46544456

#> U

#> 17831783

#> U

#> 84168812

#> U

#> 34663664

#> U

#> 71298119

#> U

#> 21792179

#> U

#> 2476937

#> U

#> 77238020

#> U

#> 64366832

#> U

#> 35653664

#> U

#> 53475149

#> U

#> 11891288

#> U

#> 45554159

#> U

#> 31693268

#> U

#> 81198317

#> U

#> 18822080

#> U

#> 5644793

#> U

#> 88128812

#> U

#> 23772179

#> U

#> 34663862

#> U

#> 47534753

#> U

#> 937955

#> U

#> 42583961

#> U

#> 67336931

#> U

#> 36647624

#> U

#> 80208515

#> U

#> 9371189

#> U

#> 42583466

#> U

#> 50501189

#> U

#> 44564456

#> U

#> 68326337

#> U

#> 991793

#> U

#> 21791783

#> U

#> 89118911

#> U

#> 87139010

#> U

#> 937937

#> U

#> 991298

#> U

#> 28722971

#> U

#> 8911937

#> U

#> 68327129

#> U

#> 61396040

#> U

#> 937946

#> U

#> 11891090

#> U

#> 69316931

#> U

#> 33672278

#> U

#> 86148713

#> U

#> 48528317

#> U

#> 34663169

#> U

#> 19811090

#> U

#> 41593763

#> U

#> 69317723

#> U

#> 946946

#> U

#> 85158218

#> U

#> 20802674

#> U

#> 84168317

#> U

#> 8920100

#> U

#> 3268496

#> U

#> 4962278

#> U

#> 928919

#> U

#> 48522674

#> U

#> 56444357

#> U

#> 12881387

#> U

#> 1000919

#> U

#> 63376634

#> U

#> 23771288

#> U

#> 12881486

#> U

#> 38628614

#> U

#> 73278020

#> U

#> 8713937

#> U

#> 33674357

#> U

#> 65355545

#> U

#> 9010937

#> U

#> 65358020

#> U

#> 8921189

#> U

#> 77238317

#> U

#> 8922179

#> U

#> 4159991

#> U

#> 84167723

#> U

#> 26742179

#> U

#> 57434753

#> U

#> 39613565

#> U

#> 10901288

#> U

#> 1387892

#> U

#> 71297723

#> U

#> 12881981

#> U

#> 15851189

#> U

#> 60408416

#> U

#> 15851090

#> U

#> 79218218

#> U

#> 16841882

#> U

#> 49514951

#> U

#> 80207723

#> U

#> 8911919

#> U

#> 19812575

#> U

#> 26741783

#> U

#> 1882793

#> U

#> 36642575

#> U

#> 17831783

#> U

#> 496397

#> U

#> 73276832

#> U

#> 59415050

#> U

#> 82188614

#> U

#> 964955

#> U

#> 86149010

#> U

#> 14862278

#> U

#> 9468911

#> U

#> 11892179

#> U

#> 33673466

#> U

#> 991955

#> U

#> 13871882

#> U

#> 45554654

#> U

#> 991955

#> U

#> 9288812

#> U

#> 50504456

#> U

#> 57433169

#> U

#> 8515928

#> U

#> 73272575

#> U

#> 23772080

#> U

#> 1288298

#> U

#> 53474456

#> U

#> 46545644

#> U

#> 15851684

#> U

#> 19812872

#> U

#> 29712872

#> U

#> 79218812

#> U

#> 946973

#> U

#> 19811288

#> U

#> 85158416

#> U

#> 8812937

#> U

#> 42586238

#> U

#> 33671981

#> U

#> 17831684

#> U

#> 15851486

#> U

#> 36643961

#> U

#> 75256139

#> U

#> 9558515

#> U

#> 937919

#> U

#> 9827921

#> U

#> 9378515

#> U

#> 80207624

#> U

#> 61395545

#> U

#> 982973

#> U

#> 982919

#> U